Adicionado material "Opaco" e revisado a estrutura do código

+++ ray/vec.cpp
@@ -15,7 +15,7 @@ Vec3 operator+(const Vec3& vlh, const Vec3& vrh) {
 	return Vec3(vlh.x + vrh.x, vlh.y + vrh.y, vlh.z + vrh.z);
 }
 
-Vec3 operator+(const Vec3 & vlh, int vrh) {
+Vec3 operator+(const Vec3 & vlh, float vrh) {
 	return Vec3(vlh.x + vrh, vlh.y + vrh, vlh.z + vrh);
 }
