+++ ray/vec.h
@@ -20,7 +20,7 @@ public:
 
 	friend Vec3 operator+(const Vec3& vlh, const Vec3& vrh);
 
-	friend Vec3 operator+(const Vec3& vlh, int vrh);
+	friend Vec3 operator+(const Vec3& vlh, float vrh);
 
 	friend Vec3 operator-(const Vec3& vlh, const Vec3& vrh);
